\name{convert_msgf_output_to_msnid}
\alias{convert_msgf_output_to_msnid}
%- Also NEED an '\alias' for EACH other topic documented here.

\title{
Convert MS-GF+ output to MSnID
}

\description{
Converts MS-GF+ output to an MSnID object.
}

\usage{
convert_msgf_output_to_msnid(x)
}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{MS-GF+ output table}
}

\value{
An object of class \code{MSnID}
}

\author{
Vladislav A Petyuk \email{vladislav.petyuk@pnnl.gov}
}

\examples{
## Missing example
}

