\name{msnidObj}
\alias{msnidObj}
\alias{c_elegans}
\docType{data}
\title{
MSnID Object
}

\description{
MSnID object from c_elegans_A_3_1_21Apr10_Draco_10-03-04_msgfplus.mzid.gz dataset
from PeptideAltas repository id PASS00308.
}

\usage{
data(c_elegans)
}

\examples{
data(c_elegans)
msnidObj
}
\keyword{datasets}
