% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSnID-methods.R
\name{read_mzIDs,MSnID-method}
\alias{read_mzIDs,MSnID-method}
\title{extra stuff
save for reuse}
\usage{
\S4method{read_mzIDs}{MSnID}(object, mzids, backend = c("mzID", "mzR"))
}
\description{
extra stuff
save for reuse
}
