% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_peptides_to_fasta_file.R
\name{map_peptides_to_fasta_file}
\alias{map_peptides_to_fasta_file}
\title{Map peptides to FASTA file}
\usage{
map_peptides_to_fasta_file(peptides, fasta, numCores = detectCores() - 1)
}
\arguments{
\item{peptides}{(character) list of clean peptide sequences}

\item{fasta}{(AAStringSet)}
}
\value{
(data.frame) with two columns for peptide and accession
}
\description{
Takes list of clean peptide sequences and FASTA file and returns a
data frame containing all matches.
}
